# backend-familleunie/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tontine Famille Unie API",
        "description": "API Laravel pour la gestion de la tontine Famille Unie",
        "contact": {
            "email": "admin@familleunie.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/meetings": {
            "get": {
                "tags": [
                    "Réunions"
                ],
                "summary": "Liste des réunions",
                "operationId": "3cc686a730c9361cbd5c74c72e1af7c7",
                "responses": {
                    "200": {
                        "description": "Liste des réunions"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Réunions"
                ],
                "summary": "Créer une réunion (ADMIN/TRESORIER)",
                "operationId": "074c4e67b80887d3284ca8e4d0d2e9a2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cycle_id",
                                    "meeting_date"
                                ],
                                "properties": {
                                    "cycle_id": {
                                        "type": "integer"
                                    },
                                    "meeting_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Réunion créée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/cycles": {
            "get": {
                "tags": [
                    "Réunions"
                ],
                "summary": "Liste des cycles",
                "operationId": "b2a09e3736f76f38c70279ac88e6f5cd",
                "responses": {
                    "200": {
                        "description": "Liste des cycles"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Réunions"
                ],
                "summary": "Créer un cycle (ADMIN)",
                "operationId": "bc5eb44890f99061cde2785a78250480",
                "responses": {
                    "201": {
                        "description": "Cycle créé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/cycles/{id}": {
            "delete": {
                "tags": [
                    "Réunions"
                ],
                "summary": "Supprimer un cycle (ADMIN)",
                "operationId": "ddf2510dd2da9c52b85b742bd9f846a6",
                "responses": {
                    "200": {
                        "description": "Cycle supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Réunions"
                ],
                "summary": "Modifier un cycle (ADMIN)",
                "operationId": "b2fc6900e5fe14c820a8274aafdde4cc",
                "responses": {
                    "200": {
                        "description": "Cycle modifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/meetings/{id}": {
            "delete": {
                "tags": [
                    "Réunions"
                ],
                "summary": "Supprimer une réunion (ADMIN/TRESORIER)",
                "operationId": "f2b5763066fd0526b6bcd538ff1c94c5",
                "responses": {
                    "200": {
                        "description": "Réunion supprimée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Entrez votre token Sanctum: Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}